Wall converts cleaned up and code moved around
