+++ Revit_Core_Engine/Convert/Physical/FromRevit/Window.cs
@@ -45,10 +45,10 @@ namespace BH.Revit.Engine.Core
             Window window = refObjects.GetValue<Window>(familyInstance.Id);
             if (window != null)
                 return window;
-            
-            PlanarSurface location = familyInstance.OpeningSurface(settings);
+
+            BH.oM.Geometry.ISurface location = familyInstance.OpeningSurface(settings);
             if (location == null)
-                BH.Engine.Reflection.Compute.RecordWarning(String.Format("Location of the window could not be retrieved from the model. A window object without location has been returned. Revit ElementId: {0}", familyInstance.Id.IntegerValue));
+                BH.Engine.Reflection.Compute.RecordError(String.Format("Location of the window could not be retrieved from the model. A window object without location has been returned. Revit ElementId: {0}", familyInstance.Id.IntegerValue));
 
             window = new Window { Location = location, Name = familyInstance.FamilyTypeFullName() };
             ElementType elementType = familyInstance.Document.GetElement(familyInstance.GetTypeId()) as ElementType;
